♻️refactor: 커밋 목록 조회와 커밋 상세 조회 로직 분리 및 파일명 변경
