+++ src/services/reports/create-report.ts
@@ -1,0 +1,23 @@
+import { CommitDetail } from "@/app/types/commit";
+import { getGithubCommitList } from "@/infra/github-api/commits/get-commit-list";
+import { getGithubCommitDetails } from "@/infra/github-api/commits/get-commit-details";
+
+const creaeteReport = async (
+  owner: string,
+  repositoryName: string,
+  branch: string,
+): Promise<CommitDetail[]> => {
+  const commits = await getGithubCommitList(owner, repositoryName, branch);
+
+  const shas = commits.map((commit) => commit.sha);
+
+  const commitDetails = await getGithubCommitDetails(
+    owner,
+    repositoryName,
+    shas,
+  );
+
+  return commitDetails;
+};
+
+export { creaeteReport };
